alterações na tela de ciclo

+++ Flashstudy_Mobile/app/src/main/java/br/com/flashstudy/flashstudy_mobile/offline/dao/CicloDao.java
@@ -25,5 +25,5 @@ public interface CicloDao {
 
     //Procura o Ciclo de um Usuario
     @Query("SELECT * FROM ciclo WHERE usuario_codigo = :codigo")
-    CicloOff getCicloByUsuario(int codigo);
+    CicloOff getCicloByUsuario(long codigo);
 }
